+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/service/ContainerInputService.java
@@ -2,8 +2,6 @@ package com.nemonotfound.nemos.inventory.sorting.service;
 
 import com.nemonotfound.nemos.inventory.sorting.models.ContainerInputContext;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.multiplayer.MultiPlayerGameMode;
-import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.ContainerInput;
 
@@ -11,8 +9,8 @@ import java.util.Optional;
 
 public class ContainerInputService {
 
-    private static final int PRIMARY_MOUSE_BUTTON = 0;
-    private static final int SECONDARY_MOUSE_BUTTON = 1;
+    public static final int PRIMARY_MOUSE_BUTTON = 0;
+    public static final int SECONDARY_MOUSE_BUTTON = 1;
 
     private static ContainerInputService INSTANCE;
 
@@ -36,19 +34,15 @@ public class ContainerInputService {
         return Optional.of(new ContainerInputContext(gameMode, player));
     }
 
-    public void performClick(AbstractContainerMenu menu, ContainerInputContext context, int slot) {
-        performContainerInput(menu, context, slot, PRIMARY_MOUSE_BUTTON, ContainerInput.PICKUP);
+    public void leftClickPickup(AbstractContainerMenu menu, ContainerInputContext context, int slot) {
+        pickup(menu, context, slot, PRIMARY_MOUSE_BUTTON);
     }
 
-    public void performRightClick(AbstractContainerMenu menu, ContainerInputContext context, int slot) {
-        performContainerInput(menu, context, slot, SECONDARY_MOUSE_BUTTON, ContainerInput.PICKUP);
+    public void rightCLickPickup(AbstractContainerMenu menu, ContainerInputContext context, int slot) {
+        pickup(menu, context, slot, SECONDARY_MOUSE_BUTTON);
     }
 
-    public void performPickup(int containerId, int slot, int mouseButton, MultiPlayerGameMode gameMode, LocalPlayer player) {
-        gameMode.handleContainerInput(containerId, slot, mouseButton, ContainerInput.PICKUP, player);
-    }
-
-    private void performContainerInput(AbstractContainerMenu menu, ContainerInputContext context, int slot, int mouseButton, ContainerInput containerInput) {
-        context.gameMode().handleContainerInput(menu.containerId, slot, mouseButton, containerInput, context.player());
+    public void pickup(AbstractContainerMenu menu, ContainerInputContext context, int slot, int mouseButton) {
+        context.gameMode().handleContainerInput(menu.containerId, slot, mouseButton, ContainerInput.PICKUP, context.player());
     }
 }