+++ Changelog.md
@@ -1,4 +1,7 @@
 # Changelog v1.21.1
 
+## Fixes
+- Wrong button position after recipe book was opened, because someone (I would never) accidentally deleted the update logic
+
 ## Changes
 - Slightly improved sorting performance